Alfa Tesla Shell v4.1 Original
Alpha Tesla v4.1 php shell is a constantly updated shell developed by Iranian hackers. We should not be denied, they have released a really solid shell and they are developing this alpha php shell every year by putting it on top, also it is not easily detected on the servers, it is well encrypted. Alfav4.1 shell is now stronger with the hidden base64 command page, I definitely recommend you to use it.